Make the import system more future-proof
Split the utilities from the widget and the dialog; we now have an enum
for supported applications, with a special value which means "all
applications", which is the only one which triggers showing the "Source"
column in EmpathyImportWidget. We need this to avoid being redundant in
the EmpathyAccountAssistant, as we already choose there the application
from which to import.