Don't assume that 'subscribe' exists.
This fixes bug #591468, which is a crash when you open a tab to a
contact from the user list in an IRC channel. Idle doesn't have a
'subscribe' list, so priv->subscribe is NULL, so this crashed. But if
you don't have a subscribe list, obviously you can't add the contact to
subscribe, so we can just guard this appropriately.

Reviewed-by: Guillaume Desmottes <guillaume.desmottes@collabora.co.uk>