Don't leak a ref when updating notification bubble
This meant that if a notification bubble for a tab had ever been
updated, closing the tab would not destroy the EmpathyChat, or indeed
close the Channel. New messages would be logged, but not presented to
the user. Reopening a tab to the same channel would work, but sending a
message would assert.

I'm pretty sure this fixes #588840.