Ack received messages from ourself.
Sumana Harihareswara reported that she had started a conversation with
herself, sent a message, and then tried to close the window, but
whenever she did so it reappeared.

This was because Empathy did not acknowledge "incoming" messages from
the user themself; hence, when it Close()d the channel, Gabble respawned
it, because it still had pending messages.