[empathy_builder_get_file] NULL out passed GObj ptrs if GtkBuilder file broken
If the GtkBuilder file passed to empathy_builder_get_file is broken, it
returns NULL, but doesn't touch the contents of any pointers passed to retrieve
GObjects. In several places we start using these pointers without checking
whether the file was actually setup, which causes Empathy to crash.

Ideally callers of empathy_builder_get_file() should check whether or not
the function returned correctly before using any of the passed pointers, but
in case they don't, make sure we're not passing pointers to random memory.