Remove dead code related to outstanding channels
With the move to the ChannelDispatcher channels we've requested will either not
be dispatched through HandleChannels (because we used the old way of directly
requesting it on the MC) or will get passed after we've started the request and
can thus be dispatched right away. The concept of oustanding channel is thus
obsolete.