[EmpathyVideoWidget] fix X crash when using Empathy with GTK+ c-s-w
In client-side-windows enabled GTK+, a call to GDK_WINDOW_XID() implicitly calls
gdk_window_ensure_native() so that a native X11 window exists serverside to
have an XID for.

Calling gdk_window_ensure_native() from a thread is not awesome and causes
Empathy to abort with one of several X errors. The fix is to call
GDK_WINDOW_XID() as soon as the widget is realized, before the XID is requested
from a thread (it would be neater to call gdk_window_ensure_native() here,
but that would force us to require GTK+ 2.18, which we don't want to do).

Fixes gnome bug #594890