Ensure the Contact List is moved to the current workspace when shown.
The new metacity/mutter 2.27 only moves windows to current workspace on
present() when the window is a dialog. The Contact List isn't one, so we
try to be smart and move explicitely the window's toplevel GdkWindow on
the current workspace before trying to present it.
This has been tested successfully with mutter 2.27 and metacity 2.26.

This fixes Bug 593883 - Empathy misbehaviour within gnome-shell.