utils: remove FIXME
Although still true, the comment above the function will suffice and
this can't really be changed without lots of making-async of other
places in Empathy.

I also don't think this is a problem in practice as everywhere in
Empathy we use the account manager should have a _prepare called
beforehand.

Signed-off-by: Jonny Lamb <jonnylamb@gnome.org>