accounts-dialog: don't always select the first account automatically
Sometimes it takes a while to get all account details, but getting the
account names is fairly quick. What tends to happen is that the tree
view of account names is populated immediately, but the account widget
is just grey until all details are retrieved. This time is enough to
select an account in the tree view for editing.

However, previously, once all details are retrieved, the first account
is automatically selected. This is of course sensible if the user has
not selected an account, but it's pretty annoying if you've already
selected your seventeenth jabber account and then it plonks you back
to number one.

This patch only selects the first account if there is not one already
selected.

Signed-off-by: Jonny Lamb <jonnylamb@gnome.org>