call-window: change the semantic of the "Video Preview" menu item (#601288)
Change it to "Always Show Video preview". We should always display the
preview when we are sending video. So by changing the semantic of this
menu it won't be possible to disable the preview while we are sending
video.

This mean that this menu isn't changed automatically anymore when the
preview is displayed.