chat: don't display pending messages before messages from logs
empathy_chat_set_tp_chat has a comment saying that show_pending_messages is a
no-op when calling during object construction. This now is now actually true
as we added a boolean saying if we can display pending messages or not.

This is to avoid displaying pending messages *before* the messages fetched
from logs (#603980).