empathy-chat: acknowledge pending messages on empathy_chat_messages_read() (#608979)
Empathy displays this handy number of unread messages in the title bar, but
you can't replicate this behaviour using an Observer because Empathy
acknowledges the messages immediately. This patch makes it so that Empathy
only acknowledges the messages when it decrements the unread messages count.

N.B. EmpathyTpChat currently seems to assume that no other client is going to
acknowledge messages for channels, it is handling. This seems valid enough,
just pointing it out.