Create the account even when user chooses to not create other one after
As the account is not created in impl_signal_apply() any more, we have to be
smarter in impl_signal_prepare and call account_assistant_apply_account_and_finish in more cases.

That also means that account_assistant_account_enabled_cb doesn't have to
terminate the assistant if the salut account has not been created yet.

Fixes bug #609220.