accounts-dialog: block the selection 'changed' signal while removing a row
There is no need to call accounts_dialog_model_selection_changed while removing
as we are going to call accounts_dialog_model_select_first right after which
will update the selection.

This has the nice side effect of letting the "remove" button sensitive after
the account has been removed. The callback used to be called recursively twice:
once during the removal of the row and once when calling
accounts_dialog_model_select_first in accounts_dialog_update_settings.
As accounts_dialog_model_selection_changed is checking if a row is selected
*before* calling accounts_dialog_update_settings, the button was unsensitived
at the end of the first call of the callback. (#609575)