Add option to make chatroom "always urgent" (#611894)
Often a user wants to be immediately notified of posts to a chatroom even when
the post does not address them directly by name.  This patch adds a room
option to make all posts to the room "urgent", meaning that the system-specific
urgency action should be taken -- notification, window urgency hint, etc.

Two specific use cases for when one may want this:

i)  Low-traffic rooms (so that one does not have to go check the room all the
time)
ii)  Error-logging room (room to which errors from some other system(s) are
logged)

The "always urgent" option is off by default when joining a room.