Make chatroom manager retain room's "always_urgent" state
This patch adds the always_urgent property to the list of chatroom
properties that are saved by the chatroom manager for the favorite
chatrooms.  This allows Empathy to bring favorite chatrooms back
up with the same notification properties they had the last time they
were used.

This seems reasonable as a room marked as "Notify Always" will most likely
be of such a character that this would be a reasonable setting every time
the room is joined.