chatroom-manager: Save correct room name
When writing out the room name to the persistance file, the room name should
properly be gotten with the get_room function.  This bug has probably not
bitten too many users as the remote room name and the favorite chatroom name
usually are usually the same; however, if the user were to change the
favorite chatroom name then this bug would become apparent.