tp-chat: when seeing if a chat can be upgraded, don't worry about handle type
When you upgrade a gtalk 1-1 chat to a PMUC, it moves from
[ChannelType=TEXT, TargetHandleType=CONTACT] to [ChannelType=TEXT,
TargetHandleType=ROOM].

When you upgrade an MSN 1-1 chat to a multi-user chat in butterfly, it
moves from [ChannelType=TEXT, TargetHandleType=CONTACT] to
[ChannelType=TEXT, TargetHandleType=NONE].

Both of these cases have the 1-1 chat properties set as fixed
properties in a class in their connection's RCC. They both also have
the upgraded set of properties as above fixed, and
o.fd.Tp.Channel.Interface.Conference(.DRAFT) in the appropriate
Allowed_Properties array.

Empathy used to look for this said channel interface in the allowed
properties of a TargetHandleType=ROOM class. As butterfly uses
TargetHandleType=NONE, this isn't so useful, so let's not care about
the handle type to fit around both gtalk and MSN.

Signed-off-by: Jonny Lamb <jonnylamb@gnome.org>