tp-chat: remove butterfly MSN HandleType=NONE workaround
Newer versions of butterfly use the conference interface, so their 1-1
chats don't implement the group interface. This patch changes nothing
there, but now when butterfly gives a channel with
TargetHandleType=NONE, it is actually a MUC, so act like that.

Thanks to the addition of the conference interface with this change,
we can check for its presence to decide whether this is old-school or
new-school butterfly.

This commit and 13a7c40c together fix bug #612461.

Signed-off-by: Jonny Lamb <jonnylamb@gnome.org>