ship a copy of empathy_log_store_empathy_get_dir in migrate-butterfly-logs.c
Code duplication isn't cool but that allows us to migrate logs when Empathy is
built with TPL as well (#613437).

The other option was to always build log-store-empathy but it's pulling
log-store and even the log-manager. I think adding this small duplication is
ok as that's temporary code that will be dropped during the 2.31 cycle.