Show status messages in compact view (#614052)
Previously, when the contact list was in compact mode your contacts'
status mesasges were not shown. With this patch, status messages are
shown in the space to the right of the contact's name, in the same
small, grey font as in the expanded view.

Only custom messages are shown; fallback strings like _("Available") are
redundant and made the list way too hard to read, so are not shown.

This patch does show that most of my contacts are too boring to have
status messages...