account-assistant: make impl_signal_apply() behave more like impl_signal_prepare()
Fixes b.g.o#615734

current_page only seems to be == PAGE_ENTER_CREATE the first time you visit the page.
After that, it has been seen to be 4 or larger. impl_signal_prepare() currently checks
for >=, so I'm copying that, and also dealing with refs in the same way.

This patch probably probably actually hides a bug in how current_page is set, because
I can't see how it should be set >= 4 because the enum only goes up to 3. I'll look into
that now, but the patch is tested and works.