Always show avatars in "Boxes" C theme
Previously, avatars were only shown in the conversation window if they
were shown in the contact list. It's not an obvious connection — I had
to look at the source to figure it out — and also I'd argue that if
anything, you're more likely to want avatars in the conversation window
if you don't have them in the contact list, because you won't ever see
them otherwise. (Or, that's what I found. :))

I left the boolean in place, forcing it to TRUE, so that if someone
feels like it they can add a tickybox to the Theme tab in Preferences.