accounts-dialog: Ensure correct connectivity information
Move the creation of the EmpathyConnectivity object to the dialog
construction so that there is time for the asynchronous calls that
the EmpathyConnectivity makes (in ConnMan mode) to return.

Without this change the error is always reported as network connection failure
rather than the real error when built with ConnMan.

Fixes: https://bugzilla.gnome.org/show_bug.cgi?id=620474