tp-chat: Wait that all the pending messages have been added to the pending queue
EmpathyChat which is using the TpChat should be sure that all the pending
messages have been processed so it can rely on
empathy_tp_chat_get_pending_messages() to return all the pending messages.

If not, it can't check if the message it receives from the logger are actually
pending ones and so will display them twice (#622729).