Ensure contacts have a persona set
Ensure that contacts created by empathy_contact_dup_from_folks_individual()
have a persona set, so that things like custom presence messages in contact
tooltips work again. This requires that we don't set the alias on the
contacts during the initial contact population period, or we'll end up with
duplicates of each contact in the list. Closes: bgo#625828