Fix crash in log_manager_got_chats_cb.
When the log window is opened through the contact list with the previous
conversations menu item, Empathy crashes due to there not being a currently
selected account when log_manager_got_chats_cb is called. This patch
adds a condition to check if there is a selected account before using it.