Don't unnecessarily remove Individuals if Individual IDs have been reused
When linking and unlinking, it's possible for the linked Individual to have
the same ID as one of its component Individuals (before linked or once
unlinked). If two such Individuals with the same ID are added and removed at
the same time (as could [only] happen during linking or unlinking),
EmpathyIndividualManager should do the Individual removals before the
additions, in order to prevent one of the Individuals getting clobbered.
Closes: bgo#627819