Add a heuristic to prefer audio- and video-capable Personas when calling
When choosing which Persona out of an Individual to start an audio or video
call to, given a choice between two Personas of equal presence, choose the
one which is capable of both audio and video calls over the one which is
capable of only one of the two. This is because clients which can do both
types of call are generally more featureful than those which can only do one
type. Helps: bgo#628338