Listen to FolksIndividual::removed in EmpathyIndividualWidget
In certain cases, a tooltip will be updated on the contact list when unlinking
a meta-contact (reproducable with a meta-contact of meta-contacts). In order
to prevent an individual table from getting stuck in the
EmpathyIndividualWidget in the tooltip, EmpathyIndividualWidget needs to
connect to FolksIndividual::removed to be updated when the Individual is
removed or replaced.